move common utils to target_utils.py
