+++ src/pubchem_target_summaries.py
@@ -10,7 +10,6 @@ separated by a newline.
 import argparse
 import json
 import os
-import re
 from typing import Tuple
 
 import pubchempy as pcp
@@ -18,6 +17,7 @@ import requests
 from tqdm import tqdm
 
 from utils.file_utils import read_aid_file
+from utils.target_utils import is_valid_uniprot_id, strip_version
 
 
 def parse_args(parser: argparse.ArgumentParser):
@@ -55,24 +55,6 @@ def get_target_type_and_id(item: dict) -> Tuple[str, str]:
             return "Pathway", id_info["other"]
         raise ValueError(f"Unrecognized target type in item: {item}")
     return None, None
-
-
-def strip_version(protein_accession: str):
-    # some accession strings include version at the end
-    # e.g., "NP_004408.1" indicates version 1 (".1")
-    # remove these versions so the accession can be used w/ PubChem API
-    pos = protein_accession.rfind(".")
-    if pos != -1:
-        return protein_accession[:pos]
-    return protein_accession
-
-
-UNIPROT_REGEX = r"[OPQ][0-9][A-Z0-9]{3}[0-9]|[A-NR-Z][0-9]([A-Z][A-Z0-9]{2}[0-9]){1,2}"
-
-
-def is_valid_uniprot_id(uniprot_candidate: str):
-    # see https://www.uniprot.org/help/accession_numbers
-    return re.match(UNIPROT_REGEX, uniprot_candidate)
 
 
 def extract_uniprot_id(json_data: dict):
